% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.changeScore.FUN}
\alias{calc.changeScore.FUN}
\title{calculation of change scores}
\usage{
calc.changeScore.FUN(Input, Timepoint.names, Groups)
}
\arguments{
\item{Input}{list of data frames with poled CTPs}

\item{Timepoint.names}{names of the test timepoints}

\item{Groups}{Groups of Treatment used for the sublists e.g. Patients and Controls}
}
\value{
list of data frames with calculated change scores
}
\description{
calculation of change scores
}
