% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.combined.Zscore.FUN}
\alias{calc.combined.Zscore.FUN}
\title{calculate combined Z- scores}
\usage{

  calc.combined.Zscore.FUN(Input.all = Config$CogDat$no_imp$Zscore.Baseline.Patients,
  Control.all = Config$CogDat$no_imp$Zscore.Baseline.Control,
  Timepoint.all = "T0")
}
\arguments{
\item{Input}{data frame containing RCI values for each CTP}

\item{Control}{control group}
}
\description{
calculate combined Z- scores
}
