% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.NCD.2.FUN}
\alias{calc.NCD.2.FUN}
\title{calculate NCD with two dichotom parameter}
\usage{
calc.NCD.2.FUN(Input.all, Control.all, impairment, Timepoint.names,
  CTP.names, CTP.number, NCD.names.1, NCD.names.2)
}
\arguments{
\item{Input.all}{list fo data frames containing CTPs /RCIs of patients group}

\item{Control.all}{list fo data frames containing CTPs/RCIs of control group}

\item{impairment}{list of data frames containing the information about the impairment of the patients in logicals (TRUE, FALSE NA)}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}

\item{CTP.number}{numer of used CTP test}

\item{NCD.names.1}{name of column containing the variables for functional impairment}

\item{NCD.names.2}{name of column containing the variables for cognitive impairment}
}
\value{
list of data frames comtaining NCD diagnosis for each patient
}
\description{
calculate NCD with two dichotom parameter
}
