% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.NCD.FUN}
\alias{calc.NCD.FUN}
\title{calculate NCD}
\usage{

  calc.NCD.FUN(Input.all = Config$CogDat$CTP.Poled.missForest[grep("Patients",
  names(Config$CogDat$CTP.Poled.missForest))],
  Control.all = Config$CogDat$CTP.Poled.missForest[grep("Controls",
  names(Config$CogDat$CTP.Poled.missForest))],
  impairment = Config$CogDat$CTPs.control.clean.interim[grep("Patients",
  names(Config$CogDat$CTPs.control.clean.interim))],
  Timepoint.all = "T0", CTPs.names = Config$parameter$colnames.CTPs)
}
\arguments{
\item{Input}{data frame CTPs}

\item{Control}{control group}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}
}
\description{
calculate NCD
}
