% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.Zscore.FUN}
\alias{calc.Zscore.FUN}
\title{calculate Z- scores}
\usage{
calc.Zscore.FUN(Input.all, Control.all, Timepoint.names, CTP.names)
}
\arguments{
\item{Input.all}{list of data frames with corrected change score of patients group}

\item{Control.all}{list of data frames with corrected change score of control group}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}
}
\value{
list of data frames with calculated z scores
}
\description{
calculate Z- scores
}
