% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{check.rawdata.FUN}
\alias{check.rawdata.FUN}
\title{check rawdata for consistence}
\usage{
check.rawdata.FUN(Config.colnames = colnames.Config,
  Config.table = Config$Config.Table,
  rawdata.colnames = colnames.CogDat,
  rawdata.table = Config$CogDat$Input, Timepoint.names = Timepoints,
  CTP.number = CTPs)
}
\arguments{
\item{Config.colnames}{given colnames of the CTP_information_table}

\item{Config.table}{CTP_information_table}

\item{rawdata.colnames}{given colnames of the rawdata_table}

\item{rawdata.table}{rawdata_table}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.number}{number of CTP tests}
}
\description{
check rawdata for consistence
}
