% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{cleaning.cg.FUN}
\alias{cleaning.cg.FUN}
\title{remove problematic controls group measures}
\usage{
cleaning.cg.FUN(Input, patterns = c("do not use"), CG.cleaning)
}
\arguments{
\item{Input}{list of data frames with CTPs and cleaning variable}

\item{patterns}{string pattern, which match with problematic measures}

\item{CG.cleaning}{name of variable coding for reason not to use}
}
\value{
list of data frames with problematic control group measures removed
}
\description{
remove problematic controls group measures
}
