% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{correct.changeScore.FUN}
\alias{correct.changeScore.FUN}
\title{corrected change scores}
\usage{
correct.changeScore.FUN(Input, Control, Timepoint.names, CTP.names)
}
\arguments{
\item{Input}{list of data frames with uncorrected change score of patients group}

\item{Control}{list of data frames with uncorrected change score of control group}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}
}
\value{
list of data frames with corrected change scores
}
\description{
corrected change scores
}
