% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{doc.FUN}
\alias{doc.FUN}
\title{seperate documentation and reduce CTP-tables to CTP values only}
\usage{
doc.FUN(Input = Config$CogDat)
}
\arguments{
\item{Input}{data frame with CTPs and cleaning variable}
}
\value{
data frame with complete NA-rows removed
}
\description{
seperate documentation and reduce CTP-tables to CTP values only
}
