% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{grouping.subsets.FUN}
\alias{grouping.subsets.FUN}
\title{grouping into subsets}
\usage{
grouping.subsets.FUN(Input, Timepoint.names, Controls, Patients, names.col)
}
\arguments{
\item{Input}{data frame that contains the complete data}

\item{Timepoint.names}{names of timepoints}

\item{Controls}{name of control group}

\item{Patients}{names of treatment groups}

\item{names.col}{names of all parameter}
}
\value{
list of separate data frames for each group and timepoint.
}
\description{
This function creates a list of separate data frames for each group and timepoint. By default no further arguments are necessary, as all input variables have pre-defined values.
}
