% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{random_forest.FUN}
\alias{random_forest.FUN}
\title{random forest imputation - version 1}
\usage{
random_forest.FUN(Input = Config$CogDat$CTPs.participate,
  Timepoints.names = Config$parameter$names.Timepoints, seed = 145,
  CTP.names = Config$parameter$colnames.CTPs)
}
\arguments{
\item{Input}{data frame with CTPs and cleaning variable}

\item{CTP.names}{names of CTP tests}

\item{Timepoint.names}{names of the test timepoints}
}
\value{
data frame with rendom forest imputed values
}
\description{
random forest imputation - version 1
}
