% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{rci.diagnosis.FUN}
\alias{rci.diagnosis.FUN}
\title{calculate POCD criteria 7, 8, 9 according to Rasmussen et al. (2001)}
\usage{
rci.diagnosis.FUN(RCI.All, thresh = -1.96, hit = 2, RCI.Combined,
  Timepoint.names, Diagnose = "PreCI")
}
\arguments{
\item{RCI.All}{list of data frames containing z scores for each CTP}

\item{thresh}{threshhold; Z values below this threshhold are classified as "hit"; default is -1.96}

\item{hit}{how many "hits" are necessary for single POCD diagnosis; default is 2}

\item{RCI.Combined}{list of data frames containing combined z scores for each CTP}

\item{Timepoint.names}{names of the test timepoints}

\item{Diagnose}{name of Diagnosis by a hit , e.g. PreCI or POCD}
}
\value{
List with data frames containing final POCD diagnosis
}
\description{
function for calculation POCD criteria 7, 8, 9 according to Rasmussen et al. (2001)
}
