% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{remove.na.FUN}
\alias{remove.na.FUN}
\title{remove cases with no data and deceased patients}
\usage{
remove.na.FUN(Input = Config$CogDat$CTPs.worst.perform,
  CTP.names = Config$parameter$names.CTPs, modi = "full")
}
\arguments{
\item{Input}{data frame with CTPs and cleaning variable}

\item{CTP.names}{names of CTP tests}

\item{modi}{content of removed data : "full" - all rows with complete NA?s were removed, "onlyDead" - only dead patients were removed}
}
\value{
data frame with complete NA-rows removed
}
\description{
remove cases with no data and deceased patients
}
