% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{rename.rawdata.FUN}
\alias{rename.rawdata.FUN}
\title{rename rawdata}
\usage{
rename.rawdata.FUN(Config.colnames = colnames.Config,
  Config.table = Config$Config.Table,
  rawdata.colnames = colnames.CogDat,
  rawdata.table = Config$CogDat$Input, Timepoint.names = Timepoints,
  CTP.number = CTPs, Control = Controlgroup,
  Patients = Patientsgroup, health = name.poor.health,
  dead = name.dead, assessed = name.assessed, method = name.method)
}
\arguments{
\item{Config.colnames}{given colnames of the CTP_information_table}

\item{Config.table}{CTP_information_table}

\item{rawdata.colnames}{given colnames of the rawdata_table}

\item{rawdata.table}{rawdata_table}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.number}{number of CTP tests}

\item{Control}{name of Controlgroup}

\item{Patients}{names of treatment groups}

\item{health}{name of variable coding poor health condition}

\item{dead}{name of variable coding for deceased}

\item{assessed}{name of variable coding for successfull examination}
}
\description{
rename rawdata
}
