% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{replace.worst.FUN}
\alias{replace.worst.FUN}
\title{replace missings due to physical impairment with worst performance}
\usage{
replace.worst.FUN(Input, CTP.names, polarity,
  pattern = "worst performance")
}
\arguments{
\item{Input}{ist of data frames with CTPs and information about health condition}

\item{CTP.names}{names of CTP tests}

\item{polarity}{information about the polarity of the CTP tests}

\item{pattern}{string pattern, which match with problematic measures}
}
\value{
list of data frames with worst performanced imputated values
}
\description{
replace missings due to physical impairment with worst performance
}
