% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{transformation.FUN}
\alias{transformation.FUN}
\title{logarythmic transformation of CTP data}
\usage{
transformation.FUN(Input, CTP.names, transformation)
}
\arguments{
\item{Input}{list of data frames containing the CTP information}

\item{CTP.names}{names of CTP tests}

\item{transformation}{info which columns should be transform}
}
\value{
list of dataframes with logarythmic transformed CTP values
}
\description{
logarythmic transformation of CTP data
}
