% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.changeScore.FUN}
\alias{calc.changeScore.FUN}
\title{calculation of change scores}
\usage{
calc.changeScore.FUN(Input = Config$CogDat$CTP.Poled.missForest,
  Timepoint.names = Config$parameter$names.Timepoints,
  Groups = names(Config$parameter$names.Group))
}
\arguments{
\item{Input}{data frame with poled CTPs}

\item{Timepoint.names}{names of the test timepoints}

\item{Groups}{Groups of Treatment used for the sublists e.g. Patients and Controls}
}
\description{
calculation of change scores
}
