% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.combined.Zscore.FUN}
\alias{calc.combined.Zscore.FUN}
\title{calculate combined Z- scores}
\usage{
calc.combined.Zscore.FUN(Input.all, Control.all, Timepoint.names)
}
\arguments{
\item{Input.all}{list of data frames containing z score for each CTP of patients group}

\item{Control.all}{list of data frames containing z score for each CTP of control group}

\item{Timepoint.names}{names of the test timepoints}
}
\value{
list of data frames with combine scores
}
\description{
calculate combined Z- scores
}
