% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{calc.Zscore.FUN}
\alias{calc.Zscore.FUN}
\title{calculate Z- scores}
\usage{

  calc.Zscore.FUN(Input.all = Config$CogDat$CTP.Poled.missForest[grep("Patients.T0",
  names(Config$CogDat$CTP.Poled.missForest))],
  Control.all = Config$CogDat$CTP.Poled.missForest[grep("Controls.T0",
  names(Config$CogDat$CTP.Poled.missForest))], Timepoint.all = "T0",
  CTPs.names = Config$parameter$colnames.CTPs)
}
\arguments{
\item{Input}{data frame with uncorrected change score of patients group}

\item{Control}{control group}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}
}
\description{
calculate Z- scores
}
