% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{check.colnames.FUN}
\alias{check.colnames.FUN}
\title{check rawdata for consistence}
\usage{
check.colnames.FUN(Input, required.colnames)
}
\arguments{
\item{Input}{Input table}

\item{required.colnames}{colnames, which are required}
}
\value{
throw out an error if other colnames are needed
}
\description{
check rawdata for consistence
}
