% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{correct.changeScore.FUN}
\alias{correct.changeScore.FUN}
\title{corrected change scores}
\usage{
correct.changeScore.FUN(Input = Config$CogDat$no_imp$CS.Delta$Patients,
  Control = Config$CogDat$no_imp$CS.Delta$Controls,
  Timepoint.names = Config$parameter$names.Timepoints,
  CTP.names = Config$parameter$colnames.CTPs)
}
\arguments{
\item{Input}{data frame with uncorrected change score of patients group}

\item{Timepoint.names}{names of the test timepoints}

\item{CTP.names}{names of CTP tests}

\item{meanControl}{mean of control group, which is divided of change score}
}
\description{
corrected change scores
}
