% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{grouping.subsets.FUN}
\alias{grouping.subsets.FUN}
\title{grouping into subsets}
\usage{
grouping.subsets.FUN(Input = Config$CogDat$Table,
  Timepoints = Config$parameter$names.Timepoints,
  Controls = Config$parameter$names.Group$Control,
  Patients = Config$parameter$names.Group$Patients,
  names.col = Config$parameter$colnames)
}
\arguments{
\item{Input}{data frame that contains the complete data}

\item{Timepoints}{names of timepoints}

\item{Controls}{name of control group}

\item{Patients}{names of treatment groups}

\item{CTP.Vars}{data frame with names of CTP variables}

\item{abort.name}{names of columns containing reasons for abort}

\item{CG.cleaning.name}{names of columns containing reasons for abort due to technical or administrative problems}
}
\description{
grouping into subsets
}
