% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{random_forest.FUN}
\alias{random_forest.FUN}
\title{random forest imputation - version 1}
\usage{
random_forest.FUN(Input, seed = 145, CTP.names)
}
\arguments{
\item{Input}{list of data frames with CTPs and cleaning variable}

\item{seed}{set seed for reproducible results}

\item{CTP.names}{names of CTP tests}
}
\value{
list of data frames with rendom forest imputed values
}
\description{
random forest imputation - version 1
}
