% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_master.R
\name{rci.diagnosis.FUN}
\alias{rci.diagnosis.FUN}
\title{calculate POCD criteria 7, 8, 9 according to Rasmussen et al. (2001)}
\usage{

  rci.diagnosis.FUN(RCI.All = Config$CogDat$no_imp$Zscore.Baseline.Patients,
  thresh = -1.96, hit = 2,
  RCI.Combined = Config$CogDat$no_imp$Zscore.combined.Baseline.Patients,
  Timepoint.all = "T0", Diagnose = "PreCI")
}
\arguments{
\item{RCI.All}{data frame containing RCI values for each CTP}

\item{thresh}{threshhold; Z values below this threshhold are classified as "hit"; default is -1.96}

\item{hit}{how many "hits" are necessary for single POCD diagnosis; default is 2}

\item{RCI.Combined}{data frame containing combined RCI values for each CTP}
}
\value{
POCD.Diagnosis: List with data frames containing final POCD diagnosis
}
\description{
calculate POCD criteria 7, 8, 9 according to Rasmussen et al. (2001)
}
