% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POCD_functions_config.R
\name{replace.worst.FUN}
\alias{replace.worst.FUN}
\title{replace missings due to physical impairment with worst performance}
\usage{
replace.worst.FUN(Input = Config$CogDat$CTPs.control.clean,
  CTP.names = Config$parameter$colnames.CTPs,
  polarity = Config$parameter$PoledVars, pattern = "worst performance")
}
\arguments{
\item{Input}{data frame with CTPs and cleaning variable}

\item{CTP.names}{names of CTP tests}

\item{poor.health}{name of variable coding poor health condition}

\item{patterns}{string pattern, which match with problematic measures}
}
\value{
data frame with worst performanced imputated values
}
\description{
replace missings due to physical impairment with worst performance
}
